Move to NextTurnAutomation to be easier to edit
